pyproject.toml
[tool.poetry]
name = "cedrf"
version = "0.1.0"
description = ""
authors = ["Awakary <helen0808@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.5"
djangorestframework = "^3.15.2"
psycopg-binary = "^3.2.4"
psycopg = "^3.2.4"
python-dotenv = "^1.0.1"
drf-writable-nested = "^0.7.1"
pytest = "^8.3.4"
pytest-django = "^4.9.0"
pytest-env = "^1.1.5"
pytest-cov = "^6.0.0"
pytest-sugar = "^1.0.0"
drf-yasg = "^1.21.8"
flake8 = "^7.1.1"
django-cors-headers = "^4.6.0"
gunicorn = "^23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
django_debug_mode = true
addopts = "-p no:warnings"


